Fix bugs found by adding an on/off switch to the colony ship
Refactor network logic so that we don't manually free objects that are
being memory-managed while they still have active references.

Refactor system server to store a pointer to the system on every node,
instead of walking the tree each time. If this uses too much memory,
we can store this in an arena, but in the meantime, this massively
simplifies the APIs of a bunch of stuff, e.g. it removes the need to
pass TSystem objects everywhere, removes the need for CachedSystems,
and removes TVisibilityHelper entirely.

Fix the logic around visibility to have an explicit initialization
stage rather than relying on the first update having an empty "old"
state, since in the init state, the "old" state of the system might
not actually match the actual current state of the node. Add code to
actually allocate memory for new nodes added when we don't change the
dynasty count (luckily this was almost already implemented...).

Clarify dkNew by separating its visibility-affecting matters from its
journal-affecting matters.

Remove the food-specific hack in the dirty kinds logic.

As a result of some of those changes, nodes will actually update when
they become invisible.

Add a test to verify that restarting the servers doesn't immediately
fail. Update the other tests to correctly notice that a bunch of nodes
disappear when you first crash.